+++ docs/changelog.rst
@@ -11,6 +11,7 @@ CHANGELOG
 * Add possibility to create attachment with external link
 * Add command import hydrobiologie stations hubeau
 * Upgrade api hubeau PC quality
+* Add control type on Land module
 
 
 1.0.3 (2022-12-15)